Restructured codebase, moved functions into "meaningful" seperate files.
