+++ raven.go
@@ -1,43 +1,27 @@
 package main
 
 import (
-	"bytes"
-	"crypto/rsa"
-	"crypto/x509"
 	"encoding/base64"
-	"encoding/json"
-	"encoding/pem"
 	"flag"
 	"fmt"
-	"net"
 
 	"io/ioutil"
 	"math/rand"
-	"net/http"
 	"os"
 
-	sealedSecretPkg "github.com/bitnami-labs/sealed-secrets/pkg/apis/sealed-secrets/v1alpha1"
 	git "github.com/go-git/go-git/v5"
-	"github.com/go-git/go-git/v5/plumbing/transport"
-	gitssh "github.com/go-git/go-git/v5/plumbing/transport/ssh"
 	"github.com/hashicorp/vault/api"
-	"golang.org/x/crypto/ssh"
 
-	//"gopkg.in/src-d/go-billy.v4/memfs"
-	"github.com/go-git/go-git/v5/plumbing/object"
-	log "github.com/sirupsen/logrus"
-	yaml "gopkg.in/yaml.v2"
-	"k8s.io/api/core/v1"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	serializer "k8s.io/apimachinery/pkg/runtime/serializer"
-	k8sJson "k8s.io/apimachinery/pkg/runtime/serializer/json"
-	//"os/exec"
 	"path/filepath"
 	"reflect"
 	"strconv"
 	"strings"
-	"sync"
 	"time"
+
+	"github.com/go-git/go-git/v5/plumbing/object"
+	log "github.com/sirupsen/logrus"
+	"k8s.io/api/core/v1"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
 
 type config struct {
@@ -134,330 +118,6 @@ func Alive(slice []interface{}, val string) bool {
 }
 
 /*
-PickRipeSecrets() uses Alive() to check if we have dead secrets
-
-*/
-
-func PickRipeSecrets(PreviousKV *api.Secret, NewKV *api.Secret) (RipeSecrets []string) {
-	log.WithFields(log.Fields{
-		"previousKeys": PreviousKV.Data["keys"],
-		"newKV":        NewKV.Data["keys"],
-	}).Debug("PickRipeSecrets is starting to compare lists")
-
-	if PreviousKV.Data["keys"] == nil || NewKV.Data["keys"] == nil {
-		// we assume this is our first run so we do not know difference yet.
-		log.WithFields(log.Fields{
-			"previousKeys": PreviousKV.Data["keys"],
-			"newKV":        NewKV.Data["keys"],
-		}).Debug("PickRipeSecrets compared lists and found that either of the lists were nil")
-
-	} else if reflect.DeepEqual(PreviousKV.Data["keys"], NewKV.Data["keys"]) {
-		log.WithFields(log.Fields{
-			"previousKeys": PreviousKV.Data["keys"],
-			"newKV":        NewKV.Data["keys"],
-		}).Debug("PickRipeSecrets: Lists match.")
-	} else {
-		for _, v := range PreviousKV.Data["keys"].([]interface{}) {
-			isAlive := Alive(NewKV.Data["keys"].([]interface{}), v.(string))
-			if !isAlive {
-				log.WithFields(log.Fields{"PreviousKV.Data": PreviousKV.Data}).Debug("PickRipeSecrets: We have found a ripe secret. adding it to list of ripesecrets now.")
-				log.WithFields(log.Fields{"RipeSecret": v.(string)}).Info("PickRipeSecrets: We have found a ripe secret. adding it to list of ripesecrets now.")
-				RipeSecrets = append(RipeSecrets, v.(string))
-				log.WithFields(log.Fields{"RipeSecret": RipeSecrets}).Debug("PickRipeSecrets final list of ripe secrets")
-			}
-		}
-	}
-	return
-}
-
-/*
-HarvestRipeSecrets() checks local files based on RipeSecrets returned from PickRipeSecrets() and marks them for deletion.
-
-*/
-
-func HarvestRipeSecrets(RipeSecrets []string, clonePath string, destEnv string) {
-	if len(RipeSecrets) == 0 {
-	} else {
-
-		r, err := git.PlainOpen(clonePath)
-		if err != nil {
-			log.WithFields(log.Fields{"err": err}).Info("HarvestRipeSecrets plainopen failed")
-		}
-
-		w, err := r.Worktree()
-		if err != nil {
-			log.WithFields(log.Fields{"err": err}).Error("HarvestRipeSecrets worktree failed")
-		}
-		//Iterate ripe secrets and remove them from worktree and push changes.
-		for ripe := range RipeSecrets {
-			base := filepath.Join("declarative", destEnv, "sealedsecrets")
-			newbase := base + "/" + RipeSecrets[ripe] + ".yaml"
-			_, err = w.Remove(newbase)
-			if err != nil {
-				log.WithFields(log.Fields{"err": err}).Error("HarvestRipeSecrets worktree.Remove failed")
-			}
-			log.WithFields(log.Fields{"ripeSecret": RipeSecrets[ripe]}).Info("HarvestRipeSecrets found ripe secret. marked for deletion")
-		}
-		status, err := w.Status()
-
-		if err != nil {
-			log.WithFields(log.Fields{
-				"err": err,
-			}).Error("HarvestRipeSecret Worktree status failed")
-		}
-
-		if !status.IsClean() {
-
-			log.WithFields(log.Fields{"worktree": w, "status": status}).Debug("HarvestRipeSecret !status.IsClean() ")
-
-			commit, err := w.Commit(fmt.Sprintf("Raven removed ripe secret from git"), &git.CommitOptions{
-				Author: &object.Signature{
-					Name:  "Raven",
-					Email: "itte@tæll.no",
-					When:  time.Now(),
-				},
-			})
-
-			if strings.HasPrefix(newConfig.repoUrl, "ssh:") {
-				err = r.Push(&git.PushOptions{Auth: setSSHConfig()})
-				if err != nil {
-					panic(err)
-				}
-			} else {
-				err = r.Push(&git.PushOptions{})
-				if err != nil {
-					log.WithFields(log.Fields{"error": err}).Error("Raven gitPush error")
-				}
-				// Prints the current HEAD to verify that all worked well.
-				obj, err := r.CommitObject(commit)
-
-				if err != nil {
-					log.WithFields(log.Fields{"obj": obj}).Error("git show -s")
-				}
-				log.WithFields(log.Fields{"obj": obj}).Info("git show -s: commit")
-				genericPostWebHook()
-			}
-
-		}
-		log.WithFields(log.Fields{}).Debug("HarvestRipeSecrets done")
-	}
-}
-
-func setSSHConfig() (auth transport.AuthMethod) {
-	sshKey, err := ioutil.ReadFile("/secret/sshKey")
-	if err != nil {
-		log.WithFields(log.Fields{
-			"err": err,
-		}).Fatal("setSSHConfig: unable to read private key ")
-	}
-
-	signer, err := ssh.ParsePrivateKey(sshKey)
-	if err != nil {
-		WriteErrorToTerminationLog("setSSHConfig: unable to read private key")
-		log.WithFields(log.Fields{"err": err}).Fatal("setSSHConfig: ParsePrivateKey err")
-	}
-	hostKeyCallback := func(hostname string, remote net.Addr, key ssh.PublicKey) error {
-		return nil
-	}
-
-	auth = &gitssh.PublicKeys{User: "git", Signer: signer, HostKeyCallbackHelper: gitssh.HostKeyCallbackHelper{
-		HostKeyCallback: hostKeyCallback,
-	}}
-
-	return auth
-
-}
-
-func GitClone(LocalPath string, url string) {
-
-	cloneOptions := &git.CloneOptions{}
-
-	if strings.HasPrefix(url, "https://") {
-		//we assume a https string with creds in it. e.g. https://someuser:somepass@somehost/some/repository.git
-
-		cloneOptions = &git.CloneOptions{
-			URL:      url,
-			Progress: os.Stdout,
-		}
-
-	} else if strings.HasPrefix(url, "ssh://") {
-		//we set up config for ssh with keys. we expect ssh://somehost/some/repo.git
-
-		cloneOptions = &git.CloneOptions{
-			URL:      url,
-			Progress: os.Stdout,
-			Auth:     setSSHConfig(),
-		}
-	}
-	// we do the clone
-	log.WithFields(log.Fields{}).Debug("Raven GitClone")
-
-	remote, err := git.PlainClone(LocalPath, false, cloneOptions)
-	if err != nil {
-		log.WithFields(log.Fields{"error": err}).Debug("Raven GitClone error")
-
-	} else {
-		head, err := remote.Head()
-		if err != nil {
-			log.WithFields(log.Fields{"head": head, "error": err}).Warn("Gitclone Remote.head()")
-		}
-		log.WithFields(log.Fields{"head": head}).Debug("Raven GitClone complete")
-	}
-}
-
-func gitPush(LocalPath string, env string, url string) {
-	r, err := git.PlainOpen(LocalPath)
-	if err != nil {
-		WriteErrorToTerminationLog("plainOpen failed")
-		log.WithFields(log.Fields{"error": err}).Error("Raven PlainOpen failed")
-	}
-
-	w, err := r.Worktree()
-	if err != nil {
-		WriteErrorToTerminationLog("gitPush failed")
-		log.WithFields(log.Fields{"error": err}).Error("GitPush WorkTree error")
-	}
-
-	// Pull the latest changes from the origin remote and merge into the current branch
-	log.Debug("GitPush pulling")
-	if strings.HasPrefix(url, "ssh:") {
-		err = w.Pull(&git.PullOptions{RemoteName: "origin", Auth: setSSHConfig()})
-		if err != nil {
-			log.WithFields(log.Fields{"error": err}).Debug("Raven gitPush:Pull error")
-		}
-	} else {
-		err = w.Pull(&git.PullOptions{RemoteName: "origin"})
-		if err != nil {
-			log.WithFields(log.Fields{"error": err}).Debug("Raven gitPush:Pull error")
-		}
-	}
-
-	status, err := w.Status()
-	if err != nil {
-		log.WithFields(log.Fields{"error": err}).Error("Raven gitPush:worktree status error")
-	}
-	log.WithFields(log.Fields{"status": status}).Debug("Worktree status")
-
-	if !status.IsClean() {
-		log.WithFields(log.Fields{"isClean": status.IsClean()}).Debug("gitPush found that status is not clean, making commit with changes")
-		_, err = w.Add(".")
-		if err != nil {
-			log.WithFields(log.Fields{"error": err}).Error("Raven gitPush:worktree add error")
-		}
-
-		// We can verify the current status of the worktree using the method Status.
-
-		commit, err := w.Commit(fmt.Sprintf("Raven updated secrets in %s", env), &git.CommitOptions{
-			Author: &object.Signature{
-				Name:  "Raven",
-				Email: "itte@tæll.no",
-				When:  time.Now(),
-			},
-		})
-		if err != nil {
-			log.WithFields(log.Fields{"error": err}).Error("GitPush Worktree commit error")
-		}
-
-		// we need to set creds here if its a ssh connection,
-		if strings.HasPrefix(url, "ssh:") {
-			err = r.Push(&git.PushOptions{Auth: setSSHConfig()})
-			if err != nil {
-				panic(err)
-			}
-		} else {
-			err = r.Push(&git.PushOptions{})
-			if err != nil {
-				log.WithFields(log.Fields{"error": err}).Error("Raven gitPush error")
-			}
-			// Prints the current HEAD to verify that all worked well.
-			obj, err := r.CommitObject(commit)
-			fmt.Println("head: ", obj)
-
-			if err != nil {
-				log.WithFields(log.Fields{"obj": obj}).Error("git show -s")
-			}
-			log.WithFields(log.Fields{"obj": obj}).Info("git show -s: commit")
-			genericPostWebHook()
-		}
-
-	}
-}
-
-/*
-getallKvs parameters:
-enviroment(i.e qa??, dev??)
-*/
-
-func getAllKVs(env string, token string) (Secret *api.Secret, err error) {
-
-	client, err := client()
-	if err != nil {
-		log.WithFields(log.Fields{"error": err}).Error("getAllKVs client error")
-	}
-	url := env + "/metadata"
-
-	Secret, err = client.Logical().List(url)
-	if err != nil {
-		log.WithFields(log.Fields{"error": err}).Error("getAllKVs list error")
-	}
-	return Secret, err
-}
-
-/*
-getsingleKV() used to iterate struct from getAllKVs(), takes secretname as input, returns struct for single secret. Requires uniform data.
-*/
-
-func getSingleKV(env string, secretname string) (Secret *api.Secret) {
-	//url := vaultEndPoint + "/v1/" + env + "/data/" + secretname
-	client, err := client()
-	if err != nil {
-		log.WithFields(log.Fields{"error": err}).Error("getSingleKV client error")
-	}
-	path := fmt.Sprintf("%s/data/%s", env, secretname)
-
-	Secret, err = client.Logical().Read(path)
-	if err != nil {
-		log.WithFields(log.Fields{"error": err}).Error("getSingleKV client read error")
-	}
-	return
-
-}
-func RenewSelfToken(token string, vaultEndpoint string) {
-	client, err := client()
-	if err != nil {
-		log.WithFields(log.Fields{"error": err}).Error("RenewSelfToken client error")
-	}
-	clientToken, err := client.Auth().Token().RenewSelf(300) // renew for 5 more minutes.
-	fmt.Println(clientToken)
-	if err != nil {
-		log.WithFields(log.Fields{"error": err}).Error("client.token.renewself() client error")
-	}
-
-}
-
-/* validateSelftoken() takes token as input,
-returns false if tokens has errors or is invalid.
-*/
-
-func validateSelftoken(vaultEndPoint string, token string) (valid bool) {
-
-	client, err := client()
-	if err != nil {
-		fmt.Printf("client.ValidateSelfToken() failed: %s \n ", err)
-	}
-
-	_, err = client.Auth().Token().LookupSelf()
-	if err != nil {
-		log.WithFields(log.Fields{"error": err}).Error("validateSelfTokenlookupself failed")
-		valid = false
-		return valid
-	}
-	valid = true
-	return valid
-
-}
-
-/*
 scaffolding for k8s,
 createK8sSecret generates k8s secrets based on inputs:
 - name: name of secret
@@ -505,7 +165,7 @@ func createK8sSecret(name string, Namespace string, sourceenv string, dataFields
 
 					log.WithFields(log.Fields{"key": k, "value": v, "split": stringSplit, "datafields": dataFields.Data["data"]}).Debug("createK8sSecret: dataFields.Data[data] found base64-encoding")
 				} else {
-					log.WithFields(log.Fields{"key": k, "value": v,}).Warn ("key is not valid BASE64")
+					log.WithFields(log.Fields{"key": k, "value": v}).Warn("key is not valid BASE64")
 				}
 			} else if isDocumentationKey(newConfig.DocumentationKeys, k) {
 				Annotations[k] = v.(string)
@@ -560,134 +220,6 @@ func createK8sSecret(name string, Namespace string, sourceenv string, dataFields
 	return
 }
 
-/*
-createSealedSecret takes two arguments:
-publicKeyPath: path to PEM file.
-k8ssecret: kubernetes secret generated from createK8sSecret when iterating list of secrets.
-
-*/
-func createSealedSecret(publickeyPath string, k8ssecret *v1.Secret) (sealedSecret *sealedSecretPkg.SealedSecret) {
-	read, err := ioutil.ReadFile(publickeyPath)
-	if err != nil {
-		log.WithFields(log.Fields{"publickeyPath": publickeyPath, "error": err}).Fatal("createSealedSecret.ioutil.ReadFile: Cannot read publickeyPath")
-	}
-
-	block, _ := pem.Decode([]byte(read))
-	if block == nil {
-		log.WithFields(log.Fields{
-			"pemDecode": publickeyPath,
-		}).Fatal("createSealedSecret.Pem.Decode() failed to parse PEM block containing the public key")
-		WriteErrorToTerminationLog("failed to parse PEM block containing the public key")
-	}
-	var pub *x509.Certificate
-
-	pub, err = x509.ParseCertificate(block.Bytes)
-	if err != nil {
-		log.WithFields(log.Fields{"block.Bytes": block.Bytes, "error": err.Error()}).Fatal("createSealedSecret.Pem.Decode() failed to parse DER encoded public key: ")
-		WriteErrorToTerminationLog("failed to parse DER encoded public key: " + err.Error())
-	}
-	var codecs serializer.CodecFactory
-	rsaPublicKey, _ := pub.PublicKey.(*rsa.PublicKey)
-	sealedSecret, err = sealedSecretPkg.NewSealedSecret(codecs, rsaPublicKey, k8ssecret)
-	if err != nil {
-		log.WithFields(log.Fields{"sealedSecret": sealedSecret, "error": err.Error()}).Error("createSealedSecret.sealedSecretPkg.NewSealedSecret")
-		WriteErrorToTerminationLog("failed to parse DER encoded public key: " + err.Error())
-	}
-	// apparently we need to specifically assign these fields.
-	sealedSecret.TypeMeta = k8ssecret.TypeMeta
-	sealedSecret.ObjectMeta = k8ssecret.ObjectMeta
-	return
-}
-
-func SerializeAndWriteToFile(SealedSecret *sealedSecretPkg.SealedSecret, fullPath string) {
-	f, err := os.Create(fullPath)
-	if err != nil {
-		log.WithFields(log.Fields{"fullPath": fullPath, "error": err.Error()}).Fatal("SerializeAndWriteToFile.Os.Create")
-		WriteErrorToTerminationLog(err.Error())
-	}
-	e := k8sJson.NewYAMLSerializer(k8sJson.DefaultMetaFactory, nil, nil)
-	err = e.Encode(SealedSecret, f)
-	if err != nil {
-		log.WithFields(log.Fields{"fullPath": fullPath, "error": err.Error()}).Fatal("SerializeAndWriteToFile.e.encode")
-		WriteErrorToTerminationLog(err.Error())
-	}
-}
-
-/*
-	readSealedSecretAndCompareWithVaultStruct takes a vault KV as parameter as well as a filepointer pointing to a local yaml file.
-
-*/
-
-func readSealedSecretAndCompareWithVaultStruct(secret string, kv *api.Secret, filepointer string, secretEngine string) (NeedUpdate bool) {
-	NeedUpdate = false
-	VaultTimeStamp := kv.Data["metadata"].(map[string]interface{})["created_time"]
-
-	//grab SealedSecret file
-	data, err := ioutil.ReadFile(filepointer)
-	if err != nil {
-		WriteErrorToTerminationLog(err.Error())
-		log.WithFields(log.Fields{"filepointer": filepointer, "error": err.Error()}).Error("readSealedSecretAndCompareWithVaultStruct.ioutil.ReadFile")
-
-	}
-	//unmarshal it into a interface
-	v := make(map[string]interface{})
-	err = yaml.Unmarshal(data, &v)
-	if err != nil {
-		WriteErrorToTerminationLog(err.Error())
-		log.WithFields(log.Fields{"data": data, "v": v, "error": err.Error()}).Fatal("readSealedSecretAndCompareWithVaultStruct.YAML.Unmarshal")
-	}
-	// hacky way of getting variable
-	if _, ok := v["metadata"]; ok {
-		if !ok {
-			log.WithFields(log.Fields{"ok-status": ok}).Info("readSealedSecretAndCompareWithVaultStruct: we need a update here")
-			NeedUpdate = true
-		}
-		SealedSecretTime := v["metadata"].(map[interface{}]interface{})["annotations"].(map[interface{}]interface{})["created_time"]
-		SealedSecretSource := v["metadata"].(map[interface{}]interface{})["annotations"].(map[interface{}]interface{})["source"]
-		if VaultTimeStamp == SealedSecretTime || SealedSecretSource != secretEngine {
-			log.WithFields(log.Fields{"VaultTimeStamp": VaultTimeStamp, "SealedSecretTime": SealedSecretTime, "SealedSecretSource": SealedSecretSource, "secretEngine": secretEngine, "NeedUpdate": NeedUpdate}).Debug("readSealedSecretAndCompareWithVaultStruct either we have a match here, or secret is from another secretengine")
-			return
-		} else {
-			NeedUpdate = true
-			log.WithFields(log.Fields{"secret": secret, "vaultTimestamp": VaultTimeStamp, "SealedSecretTime": SealedSecretTime, "NeedUpdate": NeedUpdate}).Info("readSealedSecretAndCompareWithVaultStruct needUpdate")
-		}
-	}
-	return
-}
-
-/*
-getKVAndCreateSealedSecret combines several "maker-methods":
-* Get KV
-* make k8ssecret
-* return sealedsecretobject for further creation
-* return KV object in order to compare later.
-
-*/
-func getKVAndCreateSealedSecret(secretEngine string, secretName string, token string, destEnv string, pemFile string) (SealedSecret *sealedSecretPkg.SealedSecret, SingleKVFromVault *api.Secret) {
-	SingleKVFromVault = getSingleKV(secretEngine, secretName)
-	log.WithFields(log.Fields{"SingleKVFromVault": SingleKVFromVault}).Debug("getKVAndCreateSealedSecret.SingleKVFromVault")
-	k8sSecret := createK8sSecret(secretName, destEnv, secretEngine, SingleKVFromVault)
-	log.WithFields(log.Fields{"k8sSecret": k8sSecret}).Debug("getKVAndCreateSealedSecret.k8sSecret")
-	SealedSecret = createSealedSecret(pemFile, &k8sSecret)
-	log.WithFields(log.Fields{"SealedSecret": SealedSecret}).Debug("getKVAndCreateSealedSecret.SealedSecret")
-	return
-}
-
-/*
-ensurePathandreturnWritePath:
-* build stringpath
-* create path
-
-makes sure that basePath exists for SerializeAndWriteToFile, returning basePath.
-*/
-
-func ensurePathandreturnWritePath(clonePath string, destEnv string, secretName string) (basePath string) {
-	base := filepath.Join(clonePath, "declarative", destEnv, "sealedsecrets")
-	os.MkdirAll(base, os.ModePerm)
-	basePath = base + "/" + secretName + ".yaml"
-	return
-}
-
 var newConfig = config{
 	vaultEndpoint:     "",
 	secretEngine:      "",
@@ -697,81 +229,6 @@ var newConfig = config{
 	clonePath:         "",
 	repoUrl:           "",
 	DocumentationKeys: initAdditionalKeys(),
-}
-
-/*
-We need to init the client a lot so this is a helper function which returns a fresh client.
-*/
-
-func client() (*api.Client, error) {
-	config := &api.Config{
-		Address:    newConfig.vaultEndpoint,
-		HttpClient: http.DefaultClient,
-	}
-	client, err := api.NewClient(config)
-	if err != nil {
-		log.WithFields(log.Fields{"config": config, "error": err.Error()}).Fatal("client.api.newclient() failed")
-	}
-	client.SetToken(newConfig.token)
-	if err != nil {
-		log.WithFields(log.Fields{"config": config, "error": err.Error()}).Fatal("client.api.SetToken() failed")
-	}
-	return client, err
-}
-
-func genericPostWebHook() {
-	webHookUrl, iSset := os.LookupEnv("webhook_url")
-	if iSset {
-
-		reqBody, err := json.Marshal(map[string]string{
-			"vaultEndpoint": newConfig.vaultEndpoint,
-			"secretEngine":  newConfig.secretEngine,
-			"destEnv":       newConfig.destEnv,
-		})
-		resp, err := http.Post(webHookUrl,
-			"application/json",
-			bytes.NewBuffer(reqBody))
-
-		if err != nil {
-			print(err)
-		}
-		defer resp.Body.Close()
-		body, err := ioutil.ReadAll(resp.Body)
-		if err != nil {
-			print(err)
-		}
-		fmt.Println(string(body))
-
-	}
-}
-
-func forceRefresh(wg *sync.WaitGroup) {
-	var list, err = getAllKVs(newConfig.secretEngine, newConfig.token)
-	if err != nil {
-		log.WithFields(log.Fields{"list": list, "error": err.Error()}).Warn("forceRefresh().getAllKVs failed")
-	}
-	for _, secret := range list.Data["Keys"].([]string) {
-		SealedSecret, _ := getKVAndCreateSealedSecret(newConfig.secretEngine, secret, newConfig.token, newConfig.destEnv, newConfig.pemFile)
-		newBase := ensurePathandreturnWritePath(newConfig.clonePath, newConfig.destEnv, secret)
-		SerializeAndWriteToFile(SealedSecret, newBase)
-		log.WithFields(log.Fields{"secret": secret, "newBase": newBase}).Info("forceRefresh() rewrote secret")
-	}
-	wg.Done()
-
-}
-
-func refreshHandler(w http.ResponseWriter, r *http.Request) {
-	var wg sync.WaitGroup
-	wg.Add(1)
-	go forceRefresh(&wg)
-	wg.Wait()
-	log.WithFields(log.Fields{}).Info("refreshHandler:forceRefresh() done")
-}
-
-func handleRequests() {
-
-	http.HandleFunc("/forceRefresh", refreshHandler)
-	log.Fatal(http.ListenAndServe(":1337", nil))
 }
 
 func main() {
